Move apollo and redux to @data
