+++ src/@data/reducers/devotionals.js
